requirements/dev.txt
-r base.txt

factory-boy==2.4.1
django-webtest==1.7.9
webtest==2.0.16
fabric==1.10.1
django-debug-toolbar==1.6
